tributary_api/requirements/base.txt
Django==4.2.29
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-allauth==65.15.0
django-cors-headers==4.9.0
celery==5.6.2
redis==7.3.0
django-celery-beat==2.9.0
django-celery-results==2.6.0
daphne==4.2.1
channels==4.3.2
channels-redis==4.3.0
psycopg2-binary==2.9.11
gunicorn==25.1.0
sendgrid==6.12.5
python-decouple==3.8
Pillow==12.1.1
whitenoise==6.12.0
dj-database-url==3.1.2
boto3==1.38.33
